pyproject.toml
[tool.poetry]
name = "gm_trainer"
version = "0.1.0"
description = ""
authors = ["Maxwell Joslyn <maxwelljoslyn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
llm = "^0.13.1"
llm-claude-3 = "^0.3"
sqlite-utils = "^3.36"
python-ulid = "^2.4.0.post0"
prompt-toolkit = "^3.0.43"
gradio = "^4.26.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"